+++ ESFramework/UIKit/UIView+ESAdditions.m
@@ -8,7 +8,6 @@
 
 #import "UIView+ESAdditions.h"
 #import "ESHelpers.h"
-#import "UIGestureRecognizer+ESAdditions.h"
 #import "NSArray+ESAdditions.h"
 
 @implementation UIView (ESAdditions)
@@ -87,120 +86,6 @@
 
     } while (view && controller == nil);
     return controller;
-}
-
-- (UITapGestureRecognizer *)addTapGestureHandler:(void (^)(UITapGestureRecognizer *gestureRecognizer, UIView *view, CGPoint locationInView))handler
-{
-    NSParameterAssert(handler);
-    UITapGestureRecognizer *tap = [UITapGestureRecognizer recognizerWithHandler:^(UIGestureRecognizer *sender, UIGestureRecognizerState state, CGPoint locationInView) {
-        if (UIGestureRecognizerStateRecognized == state) {
-            handler((UITapGestureRecognizer *)sender, sender.view, locationInView);
-        }
-    }];
-    if (!self.isUserInteractionEnabled) {
-        self.userInteractionEnabled = YES;
-    }
-    [self addGestureRecognizer:tap];
-    return tap;
-}
-
-- (UILongPressGestureRecognizer *)addLongPressGestureHandler:(void (^)(UILongPressGestureRecognizer *gestureRecognizer, UIView *view, CGPoint locationInView))handler
-{
-    NSParameterAssert(handler);
-    UILongPressGestureRecognizer *longPress = [UILongPressGestureRecognizer recognizerWithHandler:^(UIGestureRecognizer *sender, UIGestureRecognizerState state, CGPoint locationInView) {
-        if (UIGestureRecognizerStateBegan == state) {
-            handler((UILongPressGestureRecognizer *)sender, sender.view, locationInView);
-        }
-    }];
-    if (!self.isUserInteractionEnabled) {
-        self.userInteractionEnabled = YES;
-    }
-    [self addGestureRecognizer:longPress];
-    return longPress;
-}
-
-- (NSArray<__kindof UITapGestureRecognizer *> *)allTapGestureRecognizers
-{
-    return [self.gestureRecognizers objectsPassingTest:^BOOL (__kindof UIGestureRecognizer *obj, NSUInteger idx, BOOL *stop) {
-        return [obj isKindOfClass:UITapGestureRecognizer.class];
-    }];
-}
-
-- (NSArray<__kindof UILongPressGestureRecognizer *> *)allLongPressGestureRecognizers
-{
-    return [self.gestureRecognizers objectsPassingTest:^BOOL (__kindof UIGestureRecognizer *obj, NSUInteger idx, BOOL *stop) {
-        return [obj isKindOfClass:UILongPressGestureRecognizer.class];
-    }];
-}
-
-- (NSArray<__kindof UIPanGestureRecognizer *> *)allPanGestureRecognizers
-{
-    return [self.gestureRecognizers objectsPassingTest:^BOOL (__kindof UIGestureRecognizer *obj, NSUInteger idx, BOOL *stop) {
-        return [obj isKindOfClass:UIPanGestureRecognizer.class];
-    }];
-}
-
-- (NSArray<__kindof UIPinchGestureRecognizer *> *)allPinchGestureRecognizers
-{
-    return [self.gestureRecognizers objectsPassingTest:^BOOL (__kindof UIGestureRecognizer *obj, NSUInteger idx, BOOL *stop) {
-        return [obj isKindOfClass:UIPinchGestureRecognizer.class];
-    }];
-}
-
-- (NSArray<__kindof UISwipeGestureRecognizer *> *)allSwipeGestureRecognizers
-{
-    return [self.gestureRecognizers objectsPassingTest:^BOOL (__kindof UIGestureRecognizer *obj, NSUInteger idx, BOOL *stop) {
-        return [obj isKindOfClass:UISwipeGestureRecognizer.class];
-    }];
-}
-
-- (NSArray<__kindof UIRotationGestureRecognizer *> *)allRotationGestureRecognizers
-{
-    return [self.gestureRecognizers objectsPassingTest:^BOOL (__kindof UIGestureRecognizer *obj, NSUInteger idx, BOOL *stop) {
-        return [obj isKindOfClass:UIRotationGestureRecognizer.class];
-    }];
-}
-
-- (void)removeAllTapGestureRecognizers
-{
-    [self.allTapGestureRecognizers enumerateObjectsUsingBlock:^(__kindof UITapGestureRecognizer * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-        [obj.view removeGestureRecognizer:obj];
-    }];
-}
-
-- (void)removeAllLongPressGestureRecognizers
-{
-    [self.allLongPressGestureRecognizers enumerateObjectsUsingBlock:^(__kindof UILongPressGestureRecognizer * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-        [obj.view removeGestureRecognizer:obj];
-    }];
-}
-
-- (void)removeAllPanGestureRecognizers
-{
-    [self.allPanGestureRecognizers enumerateObjectsUsingBlock:^(__kindof UIPanGestureRecognizer * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-        [obj.view removeGestureRecognizer:obj];
-    }];
-}
-
-- (void)removeAllPinchGestureRecognizers
-{
-    [self.allPinchGestureRecognizers enumerateObjectsUsingBlock:^(__kindof UIPinchGestureRecognizer * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-        [obj.view removeGestureRecognizer:obj];
-    }];
-}
-
-- (void)removeAllSwipeGestureRecognizers
-{
-    [self.allSwipeGestureRecognizers enumerateObjectsUsingBlock:^(__kindof UISwipeGestureRecognizer * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-        [obj.view removeGestureRecognizer:obj];
-    }];
-}
-
-- (void)removeAllRotationGestureRecognizers
-{
-    [self.allRotationGestureRecognizers enumerateObjectsUsingBlock:^(__kindof UIRotationGestureRecognizer * _Nonnull obj, NSUInteger idx, BOOL * _Nonnull stop) {
-        [obj.view removeGestureRecognizer:obj];
-    }];
 }
 
 - (void)setMaskLayerByRoundingCorners:(UIRectCorner)corners cornerRadii:(CGSize)cornerRadii